zcash_transparent

Module bundle

Structs representing the components within Zcash transactions.

Modules

testing

Structs

Authorized

Bundle

EffectsOnly

Marker type for a bundle that contains no authorizing data, and the necessary input information for creating sighashes.

OutPoint

TxIn

TxOut

Traits

Authorization

MapAuth