zcash_transparent::pczt

Struct Bundle

pub struct Bundle { /* private fields */ }

PCZT fields that are specific to producing the transaction’s transparent bundle (if any).

This struct is for representing Sapling in a partially-created transaction. If you have a fully-created transaction, use the regular Bundle struct.

Implementations

impl Bundle

pub fn parse( inputs: Vec<Input>, outputs: Vec<Output>, ) -> Result<Self, ParseError>

Parses a PCZT bundle from its component parts.

impl Bundle

pub fn update_with<F>(&mut self, f: F) -> Result<(), UpdaterError>
where F: FnOnce(Updater<'_>) -> Result<(), UpdaterError>,

Updates the bundle with information provided in the given closure.

impl Bundle

pub fn finalize_spends(&mut self) -> Result<(), SpendFinalizerError>

Finalizes the spends for this bundle.

impl Bundle

pub fn extract_effects( &self, ) -> Result<Option<Bundle<EffectsOnly>>, TxExtractorError>

Extracts the effects of this PCZT bundle as a regular Bundle.

This is used by the Signer role to produce the transaction sighash.

pub fn extract(self) -> Result<Option<Bundle<Unbound>>, TxExtractorError>

Extracts a fully authorized regular Bundle from this PCZT bundle.

This is used by the Transaction Extractor role to produce the final transaction.

impl Bundle

pub fn inputs(&self) -> &Vec<Input>

The transparent inputs in this bundle.

Entries are added by the Constructor, and modified by an Updater, IO Finalizer, Signer, Combiner, or Spend Finalizer.

pub fn outputs(&self) -> &Vec<Output>

The transparent outputs in this bundle.

Entries are added by the Constructor, and modified by an Updater, IO Finalizer, Signer, Combiner, or Spend Finalizer.

impl Bundle

pub fn inputs_mut(&mut self) -> &mut [Input]

Returns a mutable reference to the inputs in this bundle.

This is used by Signers to apply signatures with Input::sign.