zcash_transparent
Module sighash

Structs

SighashType

A ZIP 244 sighash type.

SignableInput

A transparent input that is signable because we know its value and script_pubkey.

Constants

SIGHASH_ALL

SIGHASH_ANYONECANPAY

SIGHASH_MASK

SIGHASH_NONE

SIGHASH_SINGLE

Traits

TransparentAuthorizingContext

Additional context that is needed to compute signature hashes for transactions that include transparent inputs or outputs.