zcash_transparent::sighash

Trait TransparentAuthorizingContext

pub trait TransparentAuthorizingContext: Authorization {
    // Required methods
    fn input_amounts(&self) -> Vec<Zatoshis>;
    fn input_scriptpubkeys(&self) -> Vec<Script>;
}

Additional context that is needed to compute signature hashes for transactions that include transparent inputs or outputs.

Required Methods

fn input_amounts(&self) -> Vec<Zatoshis>

Returns the list of all transparent input amounts, provided so that wallets can commit to the transparent input breakdown without requiring the full data of the previous transactions providing these inputs.

fn input_scriptpubkeys(&self) -> Vec<Script>

Returns the list of all transparent input scriptPubKeys, provided so that wallets can commit to the transparent input breakdown without requiring the full data of the previous transactions providing these inputs.