zcash_transparent::zip48

Struct FullViewingKey

pub struct FullViewingKey { /* private fields */ }

A ZIP 48 P2SH multisig full viewing key.

This provides the necessary derivation capability to view all funds controlled by a P2SH multisig account.

Implementations

impl FullViewingKey

pub fn standard( threshold: u8, key_info: Vec<AccountPubKey>, ) -> Result<Self, FullViewingKeyError>

Constructs a full viewing key for a standard ZIP 48 account.

This uses the standard wallet descriptor template. For example:

threshold == 2

key_info.len() == 3

Wallet descriptor template: "sh(sortedmulti(2,@0/**,@1/**,@2/**))"

pub fn wallet_descriptor_template(&self) -> String

Returns the BIP 388 wallet descriptor template for this full viewing key.

pub fn multipath_descriptor<P: Parameters>(&self, params: &P) -> String

Returns the BIP 389 multipath descriptor for this full viewing key.

pub fn derive_address( &self, scope: Scope, address_index: NonHardenedChildIndex, ) -> (TransparentAddress, Redeem)

Derives the scoped P2SH address for this account at the given index, along with the corresponding redeem script.